# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Black Gold e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: inStock
          in: query
          schema:
            type: ["boolean", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail with variants
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /wholesale/inquiries:
    post:
      operationId: submitWholesaleInquiry
      tags: [wholesale]
      summary: Submit B2B/wholesale inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WholesaleInquiryInput"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WholesaleInquiry"
components:
  schemas:
...
    ProductVariant:
      type: object
      required: [id, productId, nameAr, nameEn, price, sku]
      properties:
        id:
          type: integer
        productId:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        sku:
          type: string
        stock:
          type: integer
        weightGrams:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
    Product:
      type: object
      required: [id, nameAr, nameEn, descriptionAr, descriptionEn, categoryId, basePrice, imageUrl, inStock, featured, slug]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        categoryId:
          type: integer
        categoryNameAr:
          type: ["string", "null"]
        categoryNameEn:
          type: ["string", "null"]
        basePrice:
          type: number
        imageUrl:
          type: string
        gallery:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        slug:
          type: string
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        badge:
          type: ["string", "null"]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          required: [variants]
          properties:
            variants:
              type: array
              items:
                $ref: "#/components/schemas/ProductVariant"
...
    WholesaleInquiry:
      type: object
      required: [id, companyName, contactName, email, phone, productInterest, estimatedQuantity, status, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        productInterest:
          type: string
        estimatedQuantity:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    WholesaleInquiryInput:
      type: object
      required: [companyName, contactName, email, phone, productInterest, estimatedQuantity]
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        productInterest:
          type: string
        estimatedQuantity:
          type: string
        message:
          type: ["string", "null"]