# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Black Gold e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  nameAr: string;
  nameEn: string;
  descriptionAr: string;
  descriptionEn: string;
  categoryId: number;
  /** @nullable */
  categoryNameAr?: string | null;
  /** @nullable */
  categoryNameEn?: string | null;
  basePrice: number;
  imageUrl: string;
  gallery?: string[];
  inStock: boolean;
  featured: boolean;
  slug: string;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  /** @nullable */
  badge?: string | null;
}