# lib/api-zod/src/generated/types/wholesaleInquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Black Gold e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface WholesaleInquiryInput {
  companyName: string;
  contactName: string;
  email: string;
  phone: string;
  productInterest: string;
  estimatedQuantity: string;
  /** @nullable */
  message?: string | null;
}